[B] refactor: 구독 취소 이벤트 객체 디렉토리 이동
